requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg[binary]>=3.2,<4.0
pydantic>=2.8,<3.0
pydantic-settings>=2.3,<3.0
python-dotenv>=1.0,<2.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7.4,<2.0
bcrypt>=4.1,<5.0
email-validator>=2.2,<3.0
python-multipart>=0.0.9,<1.0
requests>=2.32,<3.0
streamlit>=1.40,<2.0
pandas>=2.2,<3.0
plotly>=5.24,<6.0
networkx>=3.3,<4.0
pyvis>=0.3.2,<1.0
reportlab>=4.2,<5.0
pdfplumber>=0.11,<1.0
PyPDF2>=3.0,<4.0
scikit-learn>=1.5,<2.0
sentence-transformers>=3.0,<4.0
Pillow>=10.4,<12.0
numpy>=2.0,<3.0
